+++ packages/playground/wordpress-builds/src/wordpress/get-wordpress-module-details.ts
@@ -34,7 +34,7 @@ export function getWordPressModuleDetails(wpVersion: string = "6.9"): { size: nu
 		case 'beta':
 			/** @ts-ignore */
 			return {
-				size: 54745452,
+				size: 32626999,
 				url: url_beta,
 			};